transition-qa.md: delete inaccurate info on effect of large carb entries
Deleting this, as it is not quite accurate. It is possible to get overdosed with insulin if excessive carbs are entered, although there are built-in barriers in oref0:

"But even then, Open-iAPS will only automatically give you extra insulin if your BG is rising. You could enter a million carbs in the future, and Open-iAPS would only give you SMBs once needed."

I'd suggest a rewrite based on OpenAPS docs if we need more details here.