➕ Add pip-tools dependency
The direct dependencies ("top level packages")
must now be listed in `requirements.in.

We should now use `pip-compile` to create `requirements.txt`.
    pip-compile

` pip-compile` reads `requirements.in` to generate `requirements.txt`
with the pinned direct and indirect dependencies.

The `pip-compile` program is part of the `pip-tools` package.
We need to install it first:
    python -m pip install pip-tools

IMPORTANT:
Do not to add packages/dependencies directly into `requirements.txt`
but in `requirements.in`.