👷 🗑️ Configure CD temporarily to build and deploy this feature branch
This is temporary workaround that needs to be removed when eveything works as
expected and before opening the PR.
By default this workflow deploys the `dev` branch when pushed to a repo that belongs to nightscout.
This commit updates the workflow so that it also deploys the `mkdocs-material-migration` branch
when pushed to a repo that belongs to ebouchut.

⚠️ ❗️REMOVE ME: Drop this commit before opening the PR