🚸 Open external links in a new window
See: https://github.com/JakubAndrysek/mkdocs-open-in-new-tab

💄 Materialize external links with a right upward arrow

INFO:

This does NOT work if you are accesing the site via Gitub Pages:
    https://USERNAME.github.io/trio-docs/

It works when accessing the site via https://docs.diy-trio.org/
or previewing the site locally with `mkdocs serve`.

Why?
Mkdocs-open-in-new-tab uses an absolute URL '/js/open_in_new_tab.js'
This URL won't work when deployed in a sub-folder.

✅            https://docs.diy-trio.org/js/open_in_new_tab.js
✅                http://127.0.0.1:8000/js/open_in_new_tab.js
❌    https://ericbouchut.com/trio-docs/js/open_in_new_tab.js
❌ https://ebouchut.github.io/trio-docs/js/open_in_new_tab.js