📝 Consolidate version compatibility matrix into a single file and include it in all relevant files
- Move the existing redundant "version compatibility matrix" section to a dedicated page.
- Replaced all redundant occurences of this section with an include directive
  in all the files that previously defined it.

Tutorial:

To make a section of a Markdown file available for import,
enclose it with the following HTML comments:
`<!--include-markdown-start-->` and `<!--include-markdown-end--`
like so:

    The text before the start comment will not included.
    <!--include-markdown-start-->
    Content to include is here,
    between the start and end comments.
    <!--include-markdown-end-->
    What text after the end comment will not included either.

Next, to include this section in another file, add this to the includer file:
    %
      include-markdown "build/install/mac/requirements/devices/compatibility-matrix.md"
    %}

This include Markdown feature requires the mkdocs-include-markdown-plugin
to be installed and configured (Cf. previous commit).

See: https://github.com/mondeja/mkdocs-include-markdown-plugin